Refactor: move helpers to BaseProgram
